Add files via upload
In this tutorial, we have explored the concept of Value at Risk (VaR) and learned how to calculate it using Python. We have implemented three different VaR calculation methods: Historical VaR, Parametric VaR. We have used real financial data to demonstrate the calculations and visualize the results.
VaR is a powerful tool for risk management and portfolio optimization. It provides a quantitative measure of the potential loss an investor can expect to experience under normal market conditions. However, it is important to note that VaR has its limitations and should be used in conjunction with other risk measures and risk management techniques.